+++ src/Command/Commands/Enemy/Spawn/Spawn.hh
@@ -20,7 +20,7 @@ namespace Rtype
                     Spawn() = default;
                     ~Spawn();
 
-                    void set_server(std::map<int, std::shared_ptr<Rtype::client_info>> players, int mobType, int mobID, int x, int y);
+                    void set_server(std::map<int, std::shared_ptr<Rtype::client_info>> players, int mobType, int mobID, float x, float y);
                     void set_client();
 
                     void execute_client_side();
@@ -31,8 +31,8 @@ namespace Rtype
                     std::map<int, std::shared_ptr<Rtype::client_info>> _players;
                     int _mobType;
                     int _mobID;
-                    int _x;
-                    int _y;
+                    float _x;
+                    float _y;
             };
         }
     }
