+++ src/Command/Commands/Player/Move/Move.hh
@@ -20,8 +20,8 @@ namespace Rtype
                     Move() = default;
                     ~Move();
 
-                    void set_server(std::map<int, std::shared_ptr<Rtype::client_info>> players, int playerID, int x, int y);
-                    void set_client(udp::endpoint endpoint, int x, int y);
+                    void set_server(std::map<int, std::shared_ptr<Rtype::client_info>> players, int playerID, float x, float y);
+                    void set_client(udp::endpoint endpoint, float x, float y);
 
                     void execute_client_side();
                     void execute_server_side();
@@ -31,8 +31,8 @@ namespace Rtype
                     udp::endpoint _endpoint;
                     std::map<int, std::shared_ptr<Rtype::client_info>> _players;
                     int _playerID;
-                    int _x;
-                    int _y;
+                    float _x;
+                    float _y;
             };
         }
     }
